(PAR): switch lexer to inclusive states
From jflex manual:

    A regular expression can only be matched when its associated set of
    lexical states includes the currently active lexical state of the
    scanner or if the set of associated lexical states is empty and the
    currently active lexical state is inclusive. Exclusive and inclusive
    states only differ in this one point: rules with an empty set of
    associated states.

We have only one rule with an empty set of states: it is the catch-all
`[^]` rule at the very end. It must apply to all state, hence the
change.

With exclusive states plugin had a cataleptic seizure with infinite NPEs
for unterminated byte literal like `b'\f`.