(T): do not corrupt PROPERTY_HOME_PATH in LexingTestCase
LexerTestCase is looking for the fixture files relative to IDEA home dir,
but they are situated in project source dir.

The solution to change IDEA home to project is not satisfying. If we
don't revert the change, then further tests will write some data to test
data dir, and we can't revert the change even with setUp / tearDown,
because PathManager caches System properties.

We can reimplement `doFileTest` method of `LexerTestCase`, but that's
quite a bit of copy paste.

So let's hack around the issue by cooking up dirPath relative to IDEA home.