(RES): introduce mod_decl references
Module declaration like `mod foo;` is also a reference. We cannot reuse
existing `RustReferenceImpl` here, because it is tied to paths, and we
don't have a path here. So let's create a separate
`RustModDeclReferenceImpl` for this use case.

As a collateral damage, `RustReferenceImpl` is renamed to
`RustPathReferenceImpl` and is fixed: it erroneously used text range in
the document, instead of a text range relative to a parent.