(JPS): remove jps
Turns out jps was a bad fit for the job of building Rust code. The main
reason is that there are a lot of ways to build Rust project (--release,
profiles, --test), and only in the Run Configuration can we know the
precise set of flags we need to pass to cargo.

So the current strategy is just to execute usual run configuration and
to highlight errors in console

https://github.com/intellij-rust/intellij-rust/commit/25772d2cd647184f4b0fd289bc9104e42ccd9c4a

fixes #224
fixes #232