(RES): change ModuleIndex algorithm
Module index stores a mapping from names to mod_decls. Each mod decl is
represented as an offset in the file.

Finding a super module for a given file is a two step process.

First, mod_decl candidates are found in the index based on their
names (this may give false positives).

Then, each candidate is resovled and if it resolves to the target
file (modulo usual complication with the `origianlFile` for completion),
then it is the answer.