(TOOLC): try to set up toolchain automatically
For IntelliJ IDEA users who use `import from source` this should not be
executed, because the toolchain is set up during project import.

However users of other IDEs `open` and not `import` the project, and so
the toolchain is not configured properly.

This commits tries to automatically set up the toolchain before showing
the notification.