(MISC) Fix explain filter hyperlink behavior
In IntelliJ console code, there's special case handling for the
'BrowserHyperlinkInfo' navigatable, so that setting focus to it
(when using the next occurrence / prev occurrence action) won't
automatically navigate to it (See
ConsoleViewImpl#calcNextOccurrence).

Therefore, we switch the class used to that one, so that automatically
opening web pages in the backgroud won't happen anymore as the user
jumps between occurrences in the console log. However, the user can
still click on the link directly.

Fixes #611