(INT): better base class
This adds a somewhat more safe base class for all intentions, in hope to avoid `?: return` calls in `invoke`.
I am not 100% sure that this is a splendid idea, but I like it overall, and we have to wrap API because of the
backcompat anyways.

This also removes no longer needed `runReadAction`, `runWriteAction`. This was supposed to be a separate
commit, but I've forget to actually do it, and it's to much of a hassle to disentangle them now, sorry about
that : (