CARGO: speedup module.cargoWorkspace
Attaching stdlib to workspace turned out to be a major bottleneck. One way to
fix it would be to use `CacheValuesManager` and invalidate the cache when the
workspace or `PROJECT_ROOTS` change.

But we control the code that sets the standard library, so let's just actively
update the workspace. As a side benefit, we no longer need to roundtrip via IDEA
Library class when adding stdlib to workspace