TY: Properly index pointer types
Both pointers and references are represented as RsRefLikeType.
When creating a TyFingerprint for a pointer, the fingerprint
would be based on the referenced type (T) instead of the pointer.

This would cause problems when indexing the inherent impl blocks
for const and mutable pointers.

Replace a direct access to a PSI element with a access to the stub.
This is necessary because RsStdlibResolveTest does not allow access
to the PSI during the test.