ANNOT: Handle wrong number of arguments passed to variadic functions
External C functions are allowed to be variadic.
Variadic functions take a minimum amount of arguments but passing more
arguments is valid/expected.

Our current implementation always expects the exact number of arguments
to be passed and displays an error otherwise.

Fix this behavior and show an error E0060 when not passing enough
arguments to a variadic function.

RsMethodCallExpr.expectedParamsCount() was adjusted for consistency with
RsCallExpr.expectedParamsCount() as variadic calls are only allowed
for extern C functions.