RES: Handle pat binding names overlapping with module names
When resolving paths containing "::" (e.g. modulename::func())
we are only interested in elements from the "type" namespace
not from the "value" namespace.

So skip processing of the following pat binding creating places
when resolving for types:
* if let
* while let
* for
* match arms
* lambdas

This allows us to properly handle cases where such a "if let" binding
uses the same name as a module.

Normal let bindings in blocks already do the same.