PSI: add `Members` PSI node
During the latest performance profiling, I've noticed that
`RsFunction::isInherent` is a hotstop, which really struck me as odd,
because its implementation is trivial: we look at the PSI parent (one
level, stubbed), and then look at the child of that parent (again, just
one single stubbed PSI access). Turns out, though, that finding a
specific PSI child of an impl can be very costly if it is absent,
because we'll need to look through **all** methods. So we introduce a
`Members` PSI element, to make sure that Impls and Traits have a
constantish number of children