TY: Coerce references to pointers
In places that expect a pointer it is possible to pass a reference
due to coercion.
Coercion can also remove mutability.
See https://doc.rust-lang.org/nomicon/coercions.html

&T     -> *const T
&mut T -> *mut T
&mut T -> *const T

The test in RsErrorAnnotatorTest.rs is not affected by this
because in TypeInference.coerceResolved() we currently ignore the error
where references are involved.

CC: #1753