T: Add tests for various types of method self parameters
The self parameter of a method can be written in a short hand and in
an explicit variant.
The explicit variant also supports a special case for Box<Self>
which can not be done via the implicit variant.

self      -> self: Self
&self     -> self: &Self
&mut self -> self: &mut Self
-         -> self: Box<Self>

The explicit test for "self: Self" already succeeds by accident
because the reference / mutability before and after the colon match up.