STUB: Implement a stub for RsBlock
The current code in RsFunction.innerAttrList accesses the RsBlock
which does not have a stub and thus causes the PSI to be loaded.
This is necessary because inner attributes are not children of RsFunction
but children of the RsBlock.

Other implementation of RsInnerAttributeOwner.innerAttrList
already only use stubs.

Fix this by creating a stub for RsBlock