MISC: Unify RsFunctionOwner, RsConstantOwner and RsTypeAliasOwner
The 3 Owner types are mostly the same.
The only real difference is that type aliases can not be Foreign.
That is they can not be placed in an extern { } block.

Also until now for types aliases and constants we did not
track if they were placed in an inherent or in a trait impl.

Unify them into a single RsAbstractableOwner to avoid
unnecessary code duplication.