MACRO: Expand $crate to identifier
We can't just expand `$crate` to something like `::crate_name` because
rustc always threat result of this expansion as a single identifier.
For example
```
macro_rules! foo {
    () => { bar!($crate); } // $crate consumed as a single identifier
}
macro_rules! bar {
    ($i:ident) => { use $i::*; }
}
```

Rustc figures it out by synthetic tokes (which have no text
representation). Rustc can do it this way because its macro
substitution is based on AST. But our expansion is text-based, so we
must provide something textual that can be parsed as an identifier.