ANN&REF: add annotations to leafs when highlighting leafs
Now, when we want e.g. to highlight a function name:
```rust
fn foo() {}
// ~~~ highlight this
```
we do it when Annotator.annotate is called with
leaf (identifier) element argument, not the
function element.

This is more proper usage of annotators API.
Particularly, it fixes highlights blinking
(most viewable on a dark theme).

Also:
1. fixed incorrectly highlighted "primitive"
  in this case:
  ```rust
  let mut i32 = 1;
  i32 = 2;
  ```
2. fixed literal highlighting in attributes:
  `#[foo("bar")]`