GRAM: fixed parsing of statement expressions starting with blocks
Changed FullRangeExpr and OpenRangeExpr to properly preserve the StmtMode flag stack in case of parsing error.
Fixed parsing of IndexExpr, Full/OpenRangeExpr and CastExpr when they start with a block.
Split complete/block_bin_expr.rs parser test fixture into several more specific fixtures.
Added test fixtures for newly fixed expressions, as well as for DotExpr (which are parsed as before) and CastExpr (which should error when starts with a block).