T: run build task in run configuration tests
In the beginning (before `Build View`), there wasn't a significant difference between the previous implementation and the current one.
But `Build View` introduced a separate build task that should be executed before the actual running of run configuration. Found out that build task wasn't launched in run configuration tests because of two reasons:
- `RunProfileState.execute` doesn't invoke build tasks at all
- `ExecutionManagerImpl` doesn't run build tasks in tests if `ExecutionManagerImpl.forceCompilationInTests` is `false`

So now we set `ExecutionManagerImpl.forceCompilationInTests` to true and use `ProgramRunner.execute()` (actually, `ExecutionEnvironmentBuilder.buildAndExecute()` but it calls `ProgramRunner.execute()` under the hood)