INSP: simplify lint inspection API
Decouple `RsLint` enum and highlighting type, and introduce `RsLintHighlightingType`.
It allows customizing highlighting for any lint even if you need different highlighting in different situations (see #3790 as an example) for the same lint in the same way in a single line. Previously, you had to override `RsLint#toHighlightingType` method and/or pass anonymous function