#8122: fix NPE in debugger support in IDEA
Provide default implementation of `cidr.debugger.languageSupport` extension point.
CLion relies on the existence of default implementation of this EP. And it exists in CLion.
But in IDEA with Native Debugging Support plugin, there is no such implementation by default, so we can get NPE if the platform tries to do something with non-Rust stack frame.

This change is a hack just to avoid NPE. A proper solution will be landed as a fix of https://youtrack.jetbrains.com/issue/CPP-27382