GRAM: tail expr as stmt, attrs for stmts, stub stmts
1. Parse `tail expression` as a statement (without semicolon)

Let's consider the code:

```
{

    foo();
    1
}
```

Previously it was parsed into such a tree:

```
Block
  ExprStmt
    CallExpr
  Expr
```

Now it is parsed like this:

```
Block
  ExprStmt
    CallExpr
  ExprStmt
    Expr
```

Note that previously the tail expression was parsed directly
as an Expr and now it is wrapped into ExprStmt (like any
another expression in a block).

2. Outer attributes are now attached to statements (if possible)

```
{
    #[cfg(unix)]
    foo();
}
```

```
Block
  ExprStmt
    OuterAttr
    CallExpr
```

Previously they were attached to nested expressions.

Since a tail expression is now wrapped into ExprStmt, attributes
for tail expressions are also attached to statements:

```
{
    #[cfg(unix)]
    { 0 }

    #[cfg(windows)]
    { 0 }
}
```

```
Block
  ExprStmt
    OuterAttr
    CallExpr
  ExprStmt
    OuterAttr
    CallExpr
```

2.1 Cfg attributes attached to statements are now correctly evaluated.
Particularly, now we select correct tail expression from multiple
candidates under `cfg` (see the example above).

3. Statements in constants are now stubbed:

```
const A: i32 = {
    let a = 1;
    a;
}
```

This affects nothing at the moment because we anyway skip statements
during const evaluation. But we likely want to do it in the future.

4. Statements with attributes are stubbed if there are nested items:

```
{
    #[cfg(windows)]
    let a = {
        impl Foo for Bar {}
    }
}
```

This fixes several false positives on CI.

5. Some refactoring of ExitPoint processing. Now it takes into account
cfg attributes of statements and block labels