#9077: support IDEs without `Build Tools` settings
Rider doesn't provide `Build | Build Tools` settings. As a result, `Cargo` settings cannot be added as a child of `Build Tools` and the platform adds it to as a root setting panel

The main idea is to keep `Cargo` settings under `Languages & Frameworks | Rust` if `Build | Build Tools` doesn't exist, and add `Build Tools` panel to `Cargo` settings.
To achieve it, `CargoBuildToolConfigurableProvider` and `CargoProjectConfigurableProvider` were introduced since they provide more flexible way to manage when `CargoConfigurable` should be shown and how